backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Calle Stålhem", email = "carl@stalhem.se" }
]
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "fastapi>=0.128.0",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "pydantic-settings>=2.0.0",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.0",
    "sqlmodel>=0.0.32",
    "tenacity>=9.1.4",
    "uvicorn>=0.40.0",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "ruff>=0.15.0",
]

# Ruff configuration
[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in argument defaults (FastAPI Depends pattern)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
